+++ web/ecashjs-lib/src/templates/witnesscommitment/output.js
@@ -4,7 +4,7 @@ var Buffer = require('safe-buffer').Buffer;
 var bscript = require('../../script');
 var types = require('../../types');
 var typeforce = require('typeforce');
-var OPS = require('@psf/bitcoincash-ops');
+var OPS = require('ecash-ops');
 
 var HEADER = Buffer.from('aa21a9ed', 'hex');